// web/src/interfaces/database/dataset.ts
export interface ITestingChunk {
  id: string;
  content_ltks: string;
  content: string;
  document_id: string;
  document_keyword: string;
  image_id: string;
  important_keywords: any[];
  questions?: any[];
  dataset_id: string;
  similarity: number;
  term_similarity: number;
  vector_similarity: number;
  highlight: string;
  positions: number[][];
  doc_type_kwd: string;
  document_metadata?: Record<string, any>;
}